Update README.md
Rephrase and restructure sentences so that it is more understandable and concise 
1) "
Otherwise, the user will be asked to create a password, select their gender, and enter their initial meal plan. The user will be aided with some preset meal plans. From there, the user may adjust their meal plan more precisely." ->
"The user is later then asked to select their gender and initial preset meal plan which the user may adjust more precisely later on." (the password is already mentioned above there is no need to mention it again in the second paragraph)